add basic gnomonic projection handling, reorganise
