+++ src/Pixell.jl
@@ -1,5 +1,6 @@
 module Pixell
 
+import Base: in
 using WCS
 using WCS_jll
 import WCS: AbstractWCSTransform
@@ -19,6 +20,11 @@ import LinearAlgebra: mul!, ldiv!
 import SpecialFunctions: loggamma
 
 include("enmap.jl")
+
+include("projections/arbitrary_wcs.jl")
+include("projections/car_proj.jl")
+include("projections/tan_proj.jl")
+
 include("enmap_geom.jl")
 include("enmap_ops.jl")
 include("transforms.jl")
@@ -26,8 +32,8 @@ include("plot.jl")
 include("transform_distance.jl")
 include("utils.jl")
 
-export Enmap, CarClenshawCurtis, getwcs
-export geometry, fullsky_geometry, slice_geometry, pad
+export Enmap, CarClenshawCurtis, Gnomonic, getwcs
+export geometry, fullsky_geometry, slice_geometry, pad, SkyBoundingBox
 export pix2sky, pix2sky!, sky2pix, sky2pix!, skyarea, pixareamap, pixareamap!
 export read_map, write_map
 export Alm, map2alm, alm2cl, alm2map, alm2map!
